Move index_key into ValidatorBlueprint
